+++ src/Modules/Workflows/Domain/Steps/Actions/Runners/DoActionRunner.php
@@ -44,12 +44,7 @@ class DoActionRunner implements StepRunnerInterface
             function ($step) {
                 $nodeSlug = $this->stepProcessor->getSlugFromStep($step);
 
-                $this->logger->debug(
-                    $this->stepProcessor->prepareLogMessage(
-                        'Step %1$s is a Pro feature, skipping',
-                        $nodeSlug
-                    )
-                );
+                $this->logger->debugWithArgs('Step %1$s is a Pro feature, skipping', $nodeSlug);
             }
         );
     }