+++ src/Modules/Workflows/Domain/Steps/Processors/Post.php
@@ -132,6 +132,9 @@ class Post implements StepProcessorInterface, StepPostRelatedProcessorInterface
         return $this->generalProcessor->getNodeSettings($node);
     }
 
+    /**
+     * @deprecated 4.10.0 Use the logger instead
+     */
     public function logError(string $message, int $workflowId, array $step)
     {
         $this->addErrorLogMessage($message);
@@ -142,6 +145,9 @@ class Post implements StepProcessorInterface, StepPostRelatedProcessorInterface
         $this->generalProcessor->triggerCallbackIsRunning();
     }
 
+    /**
+     * @deprecated 4.10.0 Use the logger instead
+     */
     public function prepareLogMessage(string $message, ...$args): string
     {
         return $this->generalProcessor->prepareLogMessage($message, ...$args);
@@ -152,14 +158,20 @@ class Post implements StepProcessorInterface, StepPostRelatedProcessorInterface
         $this->generalProcessor->executeSafelyWithErrorHandling($step, $callback, ...$args);
     }
 
+    /**
+     * @deprecated 4.10.0 Use the logger instead
+     */
     private function addDebugLogMessage(string $message, ...$args): void
     {
-        $this->logger->debug($this->prepareLogMessage($message, ...$args));
+        $this->logger->debugWithArgs($message, ...$args);
     }
 
+    /**
+     * @deprecated 4.10.0 Use the logger instead
+     */
     private function addErrorLogMessage(string $message, ...$args): void
     {
-        $this->logger->error($this->prepareLogMessage($message, ...$args));
+        $this->logger->errorWithArgs($message, ...$args);
     }
 
     public function setPostIdOnTriggerGlobalVariable(int $postId): void