+++ src/Modules/Workflows/Domain/Steps/Triggers/Runners/OnPostMetaChangeRunner.php
@@ -41,12 +41,7 @@ class OnPostMetaChangeRunner implements TriggerRunnerInterface
 
                 $nodeSlug = $this->stepProcessor->getSlugFromStep($step);
 
-                $this->logger->debug(
-                    $this->stepProcessor->prepareLogMessage(
-                        'Step %1$s is a Pro feature, skipping',
-                        $nodeSlug
-                    )
-                );
+                $this->logger->debugWithArgs('Step %1$s is a Pro feature, skipping', $nodeSlug);
             }
         );
     }